Upgraded rtc sa1100 driver to handle date and time registers.
The driver uses now the day register RDCR and year register
RYCR to store date and time on PXA CPUs. This could be made
optionnal in KConfig if need be. Ask kernel maintainer for
advice.
This is also the way HTCUniversal and MIO A701 handle date
and time (both based on PXA27x).