tags: honor COMPILED_SOURCE with apart output directory
When the kernel is compiled with an "O=" argument, the object files are
not necessarily in the source tree, and more probably in another tree.

In this situation, the current used check doesn't work, and
COMPILED_SOURCE tags is broken with O= builds.

This patch fixes it by looking for object files both in source tree and
potential destination tree.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>